Genetic algorithm now works.
Genetic algorithm evolution now works.  It is much slower than the
original version, probably because of the boxing of Values.  This should
be less of an issue with Genetic *programming* as fewer variables will
be used.

Note:  Because we are no evaluating in a new Context each time scores do not "monotonically" increase.  A given individual may perform worse on the next set of games.